Changed base image in Dockerfile from alpine to ubuntu
To make the action more general, the base image in Dockerfile was
changed from alpine to ubuntu. Consequently, tools like bash,
curl, git and perl will now be available in the docker container
at startup. The alternative to this is to have a minimal
base image, i.e. alpine, and let users specify needed tools
as part of the command input. For example, if curl is needed
when the action is executed, the workflow using the action
may contain

  command: 'apk curl && <some_command_which_needs_curl>'

as part of the action input block. The convenience of not having
to specify installation of tools already present in ubuntu but
not in alpine comes at the cost of an increased docker image
size.